+++ auth/Cargo.toml
@@ -10,6 +10,9 @@ tokio.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 
-supabase-auth = "0.10.13"
-jsonwebtoken = "9.3.1"
 base64 = "0.22.1"
+jsonwebtoken = "9.3.1"
+supabase-auth = "0.10.13"
+
+shared = { path = "../shared" }
+async-trait = "0.1.88"
