+++ auth/src/auth_manager/sb_manager.rs
@@ -1,0 +1,6 @@
+/// An implementation of the AuthManager trait using supabase_auth.
+use supabase_auth::models::AuthClient;
+
+pub struct SbManager {
+    client: AuthClient,
+}
