+++ auth/src/lib.rs
@@ -1,9 +1,9 @@
-use axum::{Router, middleware as axum_middleware, routing::post};
+use axum::Router;
 
+mod dto;
+mod handlers;
+mod auth_manager;
 pub mod middleware;
-pub mod services;
-
-mod models;
 
 pub fn router() -> Router {
     // // these will require full authentication to access
