--- auth/src/models.rs
@@ -1,25 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Deserialize)]
-pub struct LoginRequest {
-    pub email: String,
-}
-
-#[derive(Deserialize)]
-pub struct VerifyOtpRequest {
-    pub email: String,
-    pub token: String,
-}
-
-#[derive(Serialize)]
-pub struct AuthResponse {
-    pub access_token: String,
-    pub refresh_token: String,
-    pub expires_at: u64,
-}
-
-#[derive(Deserialize, Serialize)]
-pub struct Claims {
-    pub sub: String,    // subject, namely UUID of the user
-    pub exp: usize,     // expiration time, as UTC timestamp
-}
